Extract global logger setup (#556)
