+++ main.go
@@ -29,8 +29,6 @@ import (
 	_ "net/http/pprof" // register /debug/pprof
 	"os"
 	"os/signal"
-	"path/filepath"
-	"runtime"
 	"sort"
 	"strconv"
 	"strings"
@@ -532,25 +530,7 @@ func main() {
 
 	kingpin.Parse()
 
-	logrus.SetFormatter(&logrus.TextFormatter{
-		// Enable multiline-friendly formatter in both development (with terminal) and production (without terminal):
-		// https://github.com/sirupsen/logrus/blob/839c75faf7f98a33d445d181f3018b5c3409a45e/text_formatter.go#L176-L178
-		ForceColors:     true,
-		FullTimestamp:   true,
-		TimestampFormat: "2006-01-02T15:04:05.000-07:00",
-
-		CallerPrettyfier: func(f *runtime.Frame) (function string, file string) {
-			_, function = filepath.Split(f.Function)
-
-			// keep a single directory name as a compromise between brevity and unambiguity
-			var dir string
-			dir, file = filepath.Split(f.File)
-			dir = filepath.Base(dir)
-			file = fmt.Sprintf("%s/%s:%d", dir, file, f.Line)
-
-			return
-		},
-	})
+	logger.SetupGlobalLogger()
 	if *debugF {
 		logrus.SetLevel(logrus.DebugLevel)
 	}
